Modification authentification a terminé...

+++ src/main/java/com/goodfood/api/controller/CommentsController.java
@@ -1,11 +1,12 @@
 package com.goodfood.api.controller;
 
-import com.goodfood.api.entities.Comments;
-import com.goodfood.api.entities.ErrorLog;
-import com.goodfood.api.entities.Status;
+import com.goodfood.api.entities.*;
 import com.goodfood.api.exceptions.employees.EmployeeStatusException;
 import com.goodfood.api.services.CommentsService;
+import com.goodfood.api.services.CustomersService;
+import com.goodfood.api.services.EmployeesService;
 import com.goodfood.api.services.ErrorLogServices;
+import com.goodfood.api.servicesImpl.JwtUserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -29,7 +30,13 @@ public class CommentsController
     @Autowired
     private ErrorLogServices errorLogServices;
 
-    //private AuthenticationService authenticationService;
+    @Autowired
+    private EmployeesService employeesService;
+
+    @Autowired
+    private CustomersService customersService;
+
+
 
     // ***************
     // GET
@@ -64,12 +71,20 @@ public class CommentsController
 
     @DeleteMapping( value = "/{id}" )
     @Transactional
-    public void delete( @PathVariable( value = "id" ) int id )
+    public void delete(@PathVariable( value = "id" ) int id )
     {
-        /*Status status = authenticationService.getCurrentEmployee().getStatus();
-        generatePrivilegeErrorIf(status == Status.UTILISATEUR);*/
+        LoginDao customerUser = customersService.getCustomerByCustomerId(id);
+        LoginDao employeeUser = employeesService.getEmployeeByEmployeeId(id);
 
-        this.commentsService.deleteCommentById( id );
+        if(customerUser.getCustomerNumber().getId() == id && customerUser.getStatus() == Status.UTILISATEUR)
+        {
+            this.commentsService.deleteCommentById( id );
+        }
+
+        if(employeeUser.getStatus() == Status.ADMINISTRATEUR || employeeUser.getStatus() == Status.RESTAURATEUR ||  employeeUser.getStatus() == Status.EMPLOYEE)
+        {
+            this.commentsService.deleteCommentById( id );
+        }
     }
 
 
@@ -81,21 +96,23 @@ public class CommentsController
     @Transactional
     public ResponseEntity<Comments> updateComment(@PathVariable( value = "id" ) int id, String newContent )
     {
-        return new ResponseEntity<>( this.commentsService.updateComment( id, newContent), HttpStatus.OK );
+        LoginDao customerUser = customersService.getCustomerByCustomerId(id);
+        LoginDao employeeUser = employeesService.getEmployeeByEmployeeId(id);
+
+        if(customerUser.getCustomerNumber().getId() == id && customerUser.getStatus() == Status.UTILISATEUR ||
+                employeeUser.getStatus()== Status.EMPLOYEE || employeeUser.getStatus()== Status.RESTAURATEUR ||
+                employeeUser.getStatus()== Status.ADMINISTRATEUR)
+        {
+            return new ResponseEntity<>( this.commentsService.updateComment( id, newContent), HttpStatus.OK );
+        }
+
+        errorLogServices.recordLog(new ErrorLog( null, HttpStatus.FORBIDDEN,
+                "You have not the right authorities."));
+        throw new EmployeeStatusException();
     }
 
     // ***************
     // ERROR MANAGEMENT
     // ***************
-
-    private void generatePrivilegeErrorIf( boolean test )
-    {
-        if ( test )
-        {
-            errorLogServices.recordLog( new ErrorLog( null, HttpStatus.FORBIDDEN,
-                    "You have not the right authorities." ) );
-            throw new EmployeeStatusException();
-        }
-    }
 
 }
